+++ index.php
@@ -2,7 +2,7 @@
 
 use Spatie\CodeOutline\Parser\DirectoryParser;
 use Spatie\CodeOutline\Parser\FileParser;
-use Spatie\CodeOutline\Renderer;
+use Spatie\CodeOutline\Renderer\Renderer;
 
 require_once __DIR__.'/vendor/autoload.php';
